Extract RouteScheduleTable component from ScheduleAccordionItem
...and use it on the schedule page. By extracting the inner component from the accordion item, we decouple some implementation details of the Flowbite accordion from this piece of UI, and allow ourselves to further reuse it elsewhere if and when we need to.

This commit is a WIP.

- [ ] Show all routes/Collapse all routes buttons aren't working
- [ ] Sticky route schedule headers have a gap above them
- [ ] The datepicker appears behind the sticky headers
- [ ] All schedules should be expanded by default
